+++ src/OpenTelemetry/CHANGELOG.md
@@ -6,6 +6,12 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* [Meter.Tags](https://learn.microsoft.com/en-us/dotnet/api/system.diagnostics.metrics.meter.tags?view=net-9.0)
+  will now be considered when resolving the SDK metric to update when
+  measurements are recorded. Meters with the same name and different tags will
+  now lead to unique metrics.
+  ([#5982](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5982))
+
 ## 1.11.0-rc.1
 
 Released 2024-Dec-11